refactor(templating): rename node-only templating engine to template-renderer.node.ts
Applies the same .node.ts treatment as plugin-loader.node.ts. src/templating/index.ts
loads Node-dependent plugin template tags and builds the Liquid engine, so it can only run
in Node-capable contexts; the contextIsolated renderer already uses ui/templating/renderer-safe.
The neutral index.ts name obscured that constraint.

- src/templating/index.ts -> src/templating/template-renderer.node.ts (+ header documenting the constraint)
- update importers: runtimes/templating/templating-adapter.node.ts, templating/__tests__/liquid-compat.test.ts
- environment-utils.test.ts now imports NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME from its real source (~/common/templating/constants) instead of the node-only barrel
- update renderer-safe.ts pointer comment
